% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_paths.R
\docType{data}
\name{editions}
\alias{editions}
\title{Available editions of the criteria definitions.}
\format{
An object of class \code{list} of length 2.
}
\usage{
editions
}
\description{
A named list mapping edition date strings to their directory paths within
the installed package.
}
\keyword{datasets}
